+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.3.22] - 2018-12-14
+### Added
+- Insertion of PDF files into DOCX files.
+- Links in Playground sidebar to source code on GitHub.
+### Changed
+- Data representation of a question now includes plain attachment file
+  numbers.
+### Fixed
+- Not all thread local variables were reset before request.
+- Unicode error when exporting README.
+- Azure Blob storage missing content type.
+
 ## [0.3.21] - 2018-12-10
 ### Added
 - The `object labeler` option in `fields`.